+++ lib/components/components.dart
@@ -1,1 +1,2 @@
 export 'repository_info_tile.dart' show RepositoryInfoTile;
+export 'display_number_with_icon.dart' show DisplayNumberWithIcon;
